requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
gunicorn==21.2.0
reportlab==4.2.2
psycopg2-binary==2.9.9
